Added pin expiration support for the cli
This commit adds `expires-in` option to `ctl add`, `ctl pin add` and
corresponding API endpoints.
`expires-in` option would take duration after which the item should be
unpinned.
APIs support `expires-in` which would take duration and `expire` which
would take timestamp in nanoseconds. `expire` is when API is called
through cli. `expire` is mostly to avoid the curruption in the value of
timestamp because of back and forth coversion between duration and time.

Plus tests